requirements.txt
factorio-rcon-py==1.2.0